factor more things into handler.rs
