+++ src/handlers/handler.rs
@@ -18,8 +18,9 @@ use crate::error::Error;
 use crate::nu;
 use crate::nu::frame_to_value;
 use crate::store::{FollowOption, Frame, ReadOptions, Store};
-use crate::thread_pool::ThreadPool;
 use crate::ttl::TTL;
+use crate::nu::util::value_to_json;
+use crate::thread_pool::ThreadPool;
 
 #[derive(Clone, Debug, Serialize, Deserialize, Default)]
 pub struct Meta {
@@ -183,6 +184,191 @@ impl Handler {
         });
 
         rx.await.unwrap()
+    }
+
+    fn is_value_an_append_frame(&self, value: &Value) -> bool {
+        value
+            .as_record()
+            .ok()
+            // Ensure required fields exist
+            .filter(|record| record.get("id").is_some() && record.get("topic").is_some())
+            // Chain through meta field and handler_id check
+            .and_then(|record| record.get("meta"))
+            .and_then(|meta| meta.as_record().ok())
+            .and_then(|meta_record| meta_record.get("handler_id"))
+            .and_then(|id| id.as_str().ok())
+            .filter(|id| *id == self.id.to_string())
+            .is_some()
+    }
+
+    async fn process_frame(
+        &mut self,
+        frame: &Frame,
+        store: &Store,
+        pool: &ThreadPool,
+    ) -> Result<(), Error> {
+        eprintln!("HANDLER: {} PROCESSING: frame: {:?}", self.id, frame);
+
+        if frame.topic == format!("{}.register", &self.topic)
+            || frame.topic == format!("{}.unregister", &self.topic)
+        {
+            let _ = store
+                .append(
+                    Frame::with_topic(format!("{}.unregistered", &self.topic))
+                        .meta(serde_json::json!({
+                            "handler_id": self.id.to_string(),
+                            "frame_id": frame.id.to_string(),
+                        }))
+                        .ttl(TTL::Ephemeral)
+                        .build(),
+                )
+                .await;
+            return Err("Unregister received".into());
+        }
+
+        // Skip frames that were generated by this handler
+        if frame
+            .meta
+            .as_ref()
+            .and_then(|meta| meta.get("handler_id"))
+            .and_then(|handler_id| handler_id.as_str())
+            .filter(|handler_id| *handler_id == self.id.to_string())
+            .is_some()
+        {
+            return Ok(());
+        }
+
+        let result = self.eval_in_thread(pool, &frame).await;
+
+        match result {
+            Ok(value) => {
+                match value {
+                    Value::Nothing { .. } => (),
+                    _ => {
+                        // if the return value looks like a frame returned from a .append:
+                        // ignore it
+                        if self.is_value_an_append_frame(&value) {
+                            return Ok(());
+                        }
+
+                        // Extract ReturnOptions from handler.meta
+                        let return_options = self.meta.return_options.as_ref();
+                        let postfix = return_options
+                            .and_then(|ro| ro.postfix.as_deref())
+                            .unwrap_or(".out");
+                        let ttl = return_options.and_then(|ro| ro.ttl.clone());
+
+                        eprintln!(
+                            "HANDLER: {} RETURNING: {:?}, {} {:?}",
+                            self.id, value, postfix, ttl
+                        );
+
+                        let _ = store
+                            .append(
+                                Frame::with_topic(format!("{}{}", self.topic, postfix))
+                                    .hash(
+                                        store
+                                            .cas_insert(&value_to_json(&value).to_string())
+                                            .await
+                                            .unwrap(),
+                                    )
+                                    .meta(serde_json::json!({
+                                        "handler_id": self.id.to_string(),
+                                        "frame_id": frame.id.to_string(),
+                                    }))
+                                    .maybe_ttl(ttl)
+                                    .build(),
+                            )
+                            .await;
+                    }
+                }
+            }
+            Err(err) => {
+                eprintln!("ERROR 123: {:?}", err);
+                let _ = store
+                    .append(
+                        Frame::with_topic(format!("{}.unregister", self.topic))
+                            .meta(serde_json::json!({
+                                "handler_id": self.id.to_string(),
+                                "error": err.to_string(),
+                            }))
+                            .build(),
+                    )
+                    .await;
+                return Err(err);
+            }
+        }
+
+        Ok(())
+    }
+
+    pub async fn spawn(
+        &self,
+        store: Store,
+        pool: ThreadPool,
+    ) -> Result<tokio::sync::mpsc::Sender<bool>, Error> {
+        eprintln!("HANDLER: {:?} SPAWNING", self.meta);
+
+        let (tx_command, _rx_command) = tokio::sync::mpsc::channel(1);
+
+        let options = self.configure_read_options(&store).await;
+        let mut recver = store.read(options.clone()).await;
+
+        {
+            let store = store.clone();
+            let mut handler = self.clone();
+
+            tokio::spawn(async move {
+                while let Some(frame) = recver.recv().await {
+                    eprintln!("HANDLER: {} SEE: frame: {:?}", handler.id, frame);
+
+                    if frame.topic == format!("{}.state", handler.topic) {
+                        if let Some(hash) = &frame.hash {
+                            let content = store.cas_read(hash).await.unwrap();
+                            let json_value: serde_json::Value =
+                                serde_json::from_slice(&content).unwrap();
+                            let new_state = crate::nu::util::json_to_value(
+                                &json_value,
+                                nu_protocol::Span::unknown(),
+                            );
+                            handler.state = Some(new_state);
+                            handler.state_frame_id = Some(frame.id);
+                        }
+                        continue;
+                    }
+
+                    // Skip registration activity that occurred before this handler was registered
+                    if (frame.topic == format!("{}.register", handler.topic)
+                        || frame.topic == format!("{}.unregister", handler.topic))
+                        && frame.id <= handler.id
+                    {
+                        continue;
+                    }
+
+                    if let Err(err) = handler.process_frame(&frame, &store, &pool).await {
+                        eprintln!("HANDLER: {} ERROR: {:?}", handler.id, err);
+                        break;
+                    }
+                }
+
+                eprintln!("HANDLER: {} EXITING", handler.id);
+            });
+        }
+
+        let _ = store
+            .append(
+                Frame::with_topic(format!("{}.registered", &self.topic))
+                    .meta(serde_json::json!({
+                        "handler_id": self.id.to_string(),
+                        "tail": options.tail,
+                        "last_id": options.last_id.map(|id| id.to_string()),
+                    }))
+                    .ttl(TTL::Ephemeral)
+                    .build(),
+            )
+            .await;
+
+        Ok(tx_command)
     }
 
     fn eval(&self, frame: &crate::store::Frame) -> Result<Value, Error> {
